+++ src/layout/OwnerLayout.tsx
@@ -1,10 +1,12 @@
+import Nav from "@/components/Nav";
+import PageWrapper from "@/components/PageWrapper";
 import { Outlet } from "react-router-dom";
 
 export default function OwnerLayout() {
     return (
-        <div>
-            <h1>Owner Layout</h1>
+        <PageWrapper hasNav>
             <Outlet />
-        </div>
+            <Nav version="owner" />
+        </PageWrapper>
     );
 }
