+++ src/layout/StaffLayout.tsx
@@ -1,10 +1,12 @@
+import Nav from "@/components/Nav";
+import PageWrapper from "@/components/PageWrapper";
 import { Outlet } from "react-router-dom";
 
 export default function StaffLayout() {
     return (
-        <div>
-            <h1>Staff Layout</h1>
+        <PageWrapper hasNav>
             <Outlet />
-        </div>
+            <Nav version="staff" />
+        </PageWrapper>
     );
 }
