+++ src/router/AuthRoutes.tsx
@@ -1,10 +1,11 @@
 import { RouteObject } from "react-router-dom";
 import AuthLayout from "@/layout/AuthLayout";
-import LoginPage from "@/pages/LoginPage";
-import SignupPage from "@/pages/SignupPage";
+import LoginPage from "@/pages/auth/LoginPage";
+import SignupPage from "@/pages/auth/SignupPage";
 import KakaoRedirectPage from "@/pages/auth/KakaoRedirectPage";
 import NaverRedirectPage from "@/pages/auth/NaverRedirectPage";
 import GoogleRedirectPage from "@/pages/auth/GoogleRedirectPage";
+import BusinessCertification from "@/pages/auth/BusinessCertification";
 
 const AuthRoutes: RouteObject[] = [
     {
@@ -32,6 +33,10 @@ const AuthRoutes: RouteObject[] = [
     {
         path: "/auth/google",
         element: <GoogleRedirectPage />,
+    },
+    {
+        path: "/auth/business-certificate",
+        element: <BusinessCertification />,
     },
 ];
