+++ src/router/OwnerRoutes.tsx
@@ -1,6 +1,7 @@
 import { RouteObject } from "react-router-dom";
 import OwnerLayout from "@/layout/OwnerLayout";
-import OwnerHome from "@/pages/OwnerHome";
+import OwnerHome from "@/pages/owner/OwnerHome";
+import Test from "@/pages/owner/Test";
 
 const OwnerRoutes: RouteObject[] = [
     {
@@ -10,6 +11,10 @@ const OwnerRoutes: RouteObject[] = [
             {
                 index: true,
                 element: <OwnerHome />,
+            },
+            {
+                path: "test",
+                element: <Test />,
             },
         ],
     },
